Update
I found that most PC's auto-open file explorer when ducky is plugged in messing up the script so I added a delay that should stop this from happening. Also, it doesn't need to cd to desktop as I find this is more "discreet"  and a lot of pcs go to one drive before desktop so this has a lot of potential for error. I also cleared the recycling bin and fixed deleting the files as windows askes for confirmation